refactor: cleanup hooks and lint them
